Deprecating getAddress, made compat with hdf5 1.12
