+++ include/highfive/H5DataSpace.hpp
@@ -23,11 +23,9 @@
 #endif
 
 #include "H5Object.hpp"
+#include "bits/H5_definitions.hpp"
 
 namespace HighFive {
-
-class File;
-class DataSet;
 
 ///
 /// \brief Class representing the space (dimensions) of a dataset
